requirements.txt
requests>=2.31.0
pandas>=2.1.0
google-cloud-bigquery>=3.13.0
google-cloud-bigquery-storage>=2.24.0
pyarrow>=14.0.0
db-dtypes>=1.2.0
python-dotenv>=1.0.0
